requirements.txt
Flask
requests
fire
gunicorn
python-twitter
better_profanity